[IMPROVE] MySQL 커넥션 class 문법으로 변경
