+++ backend/.vscode/settings.json
@@ -1,7 +1,8 @@
 {
   "editor.codeActionsOnSave": {
-      "source.fixAll.eslint": true
+    "source.fixAll.eslint": true
   },
   "editor.defaultFormatter": "esbenp.prettier-vscode",
-  "editor.formatOnSave": true
+  "editor.formatOnSave": true,
+  "cSpell.words": ["officetel", "oneroom", "tworoom"]
 }
